AiModels/fitness agent/requirements.txt
streamlit>=1.28.0
openai>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
python-dotenv>=1.0.0
streamlit-option-menu>=0.3.6
streamlit-card>=0.0.5